+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added the `sprig.isSuccess` variable.
 - Added the `sprig.isError` variable.
 - Added the `sprig.message` variable.
+- Added the `sprig.flashes` variable.
 - Added the `sprig.modelId` variable.
 
 ### Changed
@@ -23,6 +24,6 @@
 
 ### Deprecated
 
-- Deprecated the `success` variable. Use `sprig.isSuccess` and `sprig.isError` instead.
-- Deprecated the `flashes` variable. Use `sprig.message` instead.
+- Deprecated the `success` variable. Use `sprig.isSuccess` or `sprig.isError` instead.
+- Deprecated the `flashes` variable. Use `sprig.message` or `sprig.flashes` instead.
 - Deprecated the `id` variable. Use `sprig.modelId` instead.